kthread: Add API to create kthreads bound to low-power CPUs
Some unimportant kthreads do not need high-performance and can safely be
bound to low-power CPUs for energy efficiency.

This commit adds an API to create kthreads bound to low-power CPUs. This
affinity cannot be altered by the kernel or userspace.

This extends the performance-critical kthread support introduced in
commit c4a0ef7e1d1e ("kernel: Add API to mark IRQs and kthreads as
performance critical").

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Panchajanya1999 <rsk52959@gmail.com>
Signed-off-by: sohamxda7 <sensoham135@gmail.com>